requirements.txt
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
newsapi-python==0.2.6
numpy==1.21.4
pandas==1.3.4
platformdirs==2.4.0
pylint==2.12.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
uk-covid19==1.2.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
